Add CloneSchema, which creates an in-memory copy of a table from an underlying schema the first time it is referenced. Yes, an in-memory database!
Add support for INSERT statements. Provider must have a rule to implement TableModificationRel appropriately for the table representation. EnumerableTableModificationRel is an example.

When reading from a table, use primitive types as close as possible to the types of the columns. Use primitive (e.g. int, not Integer) if column is not null.

Use connection's schema when resolving table names in queries.

Add a bunch of tests based on SQL Mondrian generates. They all parse/plan, and all but 3 give the right results. We now invoke PushFilterThroughJoinRule to make them faster.

Re-organize tests to allow a test to inject its own schema implementation.

Improvements to SelfPopulatingList and Pair.


Former-commit-id: aacfdc866c6cdde76d8c1b364b17fc0a3db7c46d