Add StarTable.
This is the first step in recognizing complex materializations. A "star table" is a virtual table composed of real tables joined together via many-to-one relationships. The queries that define materializations and end-user queries are canonized in terms of star tables. Matching (not done yet) will then be a matter of looking for sort, groupBy, project.

 For now, we have added a dummy schema "mat" and a dummy star table "star". Later the model will allow star tables to be defined explicitly.


Former-commit-id: ef0acca555e6d78d08ea1aa5ecc6d7b42f689544