Make it easier to define sub-classes of rule operands, and optimize operands for WindowedAggSplitterRule and RemoveTrivialProjectRule.
The new class RelOptRuleOperandChildren contains the children of an operand and the policy for dealing with them. The parameters to construct an operand are now simpler, and an operand can more easily be sub-classed. Existing methods to create operands still exist but are deprecated and will be removed in optiq-0.4.18.


Former-commit-id: 7f9cde51184c8bf5da702af1b5e0191c372e572a