[OPTIQ-367] PushFilterPastJoinRule should strengthen join type
Make RexNode.isAlwaysTrue smart enough to deal with "c IS NOT NULL" on a column declared NOT NULL.

Change how EXISTS is translated (was "c IS TRUE", now "c IS NOT NULL", where c is a marker column from the null-generating side).

Keep a dumber version of PushFilterPastJoinRule around, to keep some tests working.


Former-commit-id: 201d5917931bb5e2a455ad786a671764bc8b3833