[CALCITE-544] Implement Union in Interpreter; [CALCITE-562] Implement inner JOIN in interpreter and improve handling of scalar expressions
Also, progress towards [CALCITE-558] Add BINDABLE convention and make it the default convention for query root, but currently ENUMERABLE is still the default. You can enable BINDABLE by setting CalcitePrepareImpl.ENABLE_BINDABLE; there are 184 test failures.

New interface ArrayBindable extends Bindable. Unlike Bindable the return is guaranteed to be an Object[]. Enumerable convention expects generated code to return a Bindable, which might return custom classes or arrays, and if there is a single column, the array is optimized to the raw column value. Bindable convention assumes that relational expressions implement ArrayBindable and return arrays of column values; the arrays are not optimized if there is only one column.

New interface Scalar represents a compiled way to execute a scalar expression (or expressions) and is easily built from a RexNode or AggregateCall.

Close apache/incubator-calcite#43


Former-commit-id: 66cfb120fca58e4c809748e2bcef4d3d5a7c6b82