[CALCITE-794] Detect cycles when computing statistics
Make RelMetadataQuery methods non-static. Each active call is
registered in the RelMetadataQuery instance; it throws
CyclicMetadataException when it detects a cycle, and the caller can
catch it to return a sensible default value for the particular kind of
metadata.

Change signature of FlatLists.of(Object...) to
FlatLists.copyOf(Comparable...).

Temporarily disable all failing tests.

Add metadata for EnumerableLimit.


Former-commit-id: cabdcf44e4aec4d4ceea7f97c8c6fd9e9dbd36b1