[CALCITE-604] Tune metadata by generating a dispatcher at runtime
Currently we dispatch to metadata providers using reflection. With
this change, we generate a dispatcher class and compile using Janino.

Cache results within a metadata call. This should yield a performance
improvement when, say, a Join requires many kinds of metadata
(rowCount, averageRowSize, selectivity) and they all rely on the
rowCount of the input, and that rowCount is expensive to compute.


Former-commit-id: d14040c52f8b0bc351443cff1584e8b587378698