[CALCITE-642] Add an avatica-metrics API
Avatica-metrics is a simple API, based on what
is provided by dropwizard-metrics, and a no-op
implementation of that API. One implementation
of this API is provided using dropwizard-metrics
3.1.2. ServiceLoader is used to dynamically load
one, and only one, implementation of the metrics API.

This provides some basic instrumentation inside of Avatica
to expose interesting metrics, limited only to the server
process (although the client does have the dependency
available). No explicit metrics "sinks" are configured,
where the metrics go is an exercise for the user.

Metrics instrumentation presently includes:

* Coarse Jetty request timings
* Serialization (JSON/Protobuf) timings
* Connection and Statement cache gauges

This also includes a new module, calcite-dropwizard-hadoop-metrics2,
which acts as a bridge between dropwizard-metrics and Hadoop Metrics2.
This has no bearing on Avatica or Calcite on its own, but is a useful
tool to integrate metrics exposed by Avatica/Calcite into other metrics
stores which Dropwizard-metrics doesn't natively support (namely, Ambari).

Closes apache/calcite#189


Former-commit-id: 7e5710fc7459728a667659ea2b66895812ef839f