[CALCITE-1083] SqlNode.equalsDeep has O(n ^ 2) performance
Cause of the problem is that Litmus.fail constructs an error message string, even if the particular implementation of Litmus does not use the string. Solution is to pass a format string and arguments, which are only converted to strings if the message is used, the same as SLF4J Logger does.


Former-commit-id: 670fa73e24a7b72d6ddafd69793474083bb4b9f0