[CALCITE-1078] Detach avatica from the core calcite Maven project
The "calcite-avatica" module is still called "calcite-avatica"
but is contained in the directory "avatica/core", since avatica
is now the parent directory for all of avatica.

The present contents of the LICENSE file was incorrect, stating
that all artifacts (notably the source tarball) bundled protobuf.
This is incorrect. Protobuf was only included in the avatica (client)
jar. As such, only that artifact should have the LICENSE file with
the appropriate notice in it.


Former-commit-id: 5cee486f3ad878275ffe67e67cd8593d6bcfd62d