[CALCITE-1064] Address problematic maven-remote-resources-plugin
The current configuration of the pom with this plugin was causing
the top-level module to pull down the artifacts that were about
to be built instead of using the artifacts after they were built.

By moving the remote-resources-plugin configuration to the apache-release
profile, this defers this check from running at the start
of the build, and generates a valid DEPENDENCIES file at the
end of the build when the reactor has all of the artifacts
which were just built.


Former-commit-id: 7901eda051dc756ba5a985d1cf2fed1942459281