[CALCITE-1148] Fix RelTrait conversion (e.g. distribution, collation) (Minji Kim)
In the current calcite, trait conversion is not handled properly, e.g.
collation/distribution traits are not converted (shown by the tests).
This patch fixes this issue.

For each RelCollationTrait, introduce a new API, canConvert() which
should return true if the conversion from a trait to the other is
possible.

For each Convention, introduce two new APIs, canConvertConvention()
returns true if the convernsion is possible, and useAbstractConverters()
returns true if the trait conversion should be handle via
AbstractConverters.  By default, both functions return false.

In RelSet, when adding a new RelSubset, if the convention returns false
for useAbstractConverters(), we do not add AbstractConverters.  Even if
convention.useAbstractConverters() return true, we only add
AbstractConverters if the AbstractConverters can convert (i.e. if
RelCollationTrait.canConvert() returns true) and the conversion is
needed (i.e. if RelTrait.satisfies() returns false).

Added test cases.

Close apache/calcite#210


Former-commit-id: a3bc0d8ea3c1f320b96a4ab9641becee90a388bc