[CALCITE-1103] Correct decimal serialization with protobuf
Decimal values were incorrectly being treated as generic
"numbers" which were causing them to be serialized as a
"double" value. This lost some of the extra details on
the decimal value which caused incorrect results.

Meta.Frame was also reworked to use TypedValue for serialization
into protobuf. This, combined with a simplification of TypedValue's
serialization methods, results in an overall reduction of translation
code to/from protobufs.

Also includes a test to verify that we can still parse
Decimals (although, truncated) via the former method.


Former-commit-id: aa9db8a36046b1e812021fa1cf0973906cfffbad