[CALCITE-1202] Lock version of Jekyll for bundler
Ideally, we can lock the version of github-pages which would control
what version of Jekyll we'd be using, but all of HTML files currently
have Jekyll 2.4.0 specified in a meta-generator tag. Once we flip
over to Jekyll 3, we can also switch the jekyll version to an explicit
github-pages version instead.


Former-commit-id: b66d414661926afdb82aea71432fa18022a666e2