[CALCITE-1218] Lift try/catch to Jetty Handler to prevent exceptions propagating to Jetty
We need to make sure that we always send back an ErrorResponse to the
client when we can't execute their Request (for whatever reason).

I tried to write a test to exercise this, but ran into tons of issues
trying to do this because of the relocation of protobuf classes. This
will be alleviated by CALCITE-1224 in the future.


Former-commit-id: 9deff987ba177a6eb1ded2813e770a9233f3f2f4