[CALCITE-1243] Ensure signed fields are used in protobuf for specifying rows to return
It was found that the protobuf definitions where specifying some fields in some
messages as unsigned integers where Avatica was expecting negative values
to be meaningful. We failed to notice this in Avatica because the Java bindings
for protobuf are equivalent for signed and unsigned typed. This breaks clients
in other languages though.

Some confusing logic surrounding `maxRowCount` being used for two different
meanings (both maximum number of rows for the statement and the maximum number
of rows returned in the first frame) was also corrected since the values in
the protocol buffers were being changed.

The incorrect fields were deprecated, documented, and backwards-compatibally modified
so that older clients can continue to operate without change. Clients in languages
which do have differentiation between signedness in types will have to use the
newer version of the client and server to get this fix.


Former-commit-id: 987a257ae7cff588ac2f63518a7f75263a9bddb9