[CALCITE-1444] Add CONVERT function (Laurent Goujon)
Add support for the CONVERT scalar function as specified by the
JDBC and ODBC specifications.

Introduced a new set of symbols in SqlJdbcDataTypeName, which are used
by the parser to identify CONVERT type argument. The ODBC/JDBC type is
then converted into the correct type when converting the function to a
CAST operator.

Close apache/calcite#310


Former-commit-id: 9266d8ee224cac6013ca799dbb8cdc5a2054b657