[CALCITE-1551] Preserve alias in RelBuilder.project (Jess Balint)
This is accomplished by changing the structure of the Frame in
RelBuilder to include alias and field information for all fields in the
RelNode irrespective of their origin. Rel aliases also preserved on
group keys through aggregate operations.

In PigRelBuilder, we retain aggregate aliases, because Pig Latin
relies upon them.

Close apache/calcite#340


Former-commit-id: 31d68f71b01f016625a1def800c283820c600840