Concluding [CALCITE-1555] Enable Join.isValid
This change corrects the nullability of RexInputRefs used in a Join's
ON clause. Previously, a namespace used in an outer join was made
nullable, but that was incorrect for the ON clause. To fix, we add
class ScopeChild and record that the namespace is nullable when used
from afar, but we don't alter the row type of the namespace.


Former-commit-id: 70b8325130bc6be953e5a9a0ce99f9e0c383268b