[CALCITE-1616] Data profiler
Profiler does not work on JDK 7, because it relies on
com.yahoo.datasketches:sketches-core:0.9.0, which needs JDK 8 or
higher. Therefore some tests are disabled on JDK 7.

Change interface LatticeStatisticProvider to estimate joint rather
than singleton cardinality.

Bind LatticeStatisticProvider to a Lattice, and create
LatticeStatisticProvider.Factory to do the binding.

Keep results only if they are sufficiently "surprising" (informative).
The "surprise queue" is FIFO and holds the top N accepted values,
after a warm-up period

PartiallyOrderedSet: find "hypothetical" parents and children of
elements not in the set.

Former-commit-id: dad581868b815510389f61936b0c583be93d5dc5