Add test case for [CALCITE-1769] Push filter involving cast to numeric down to Druid
The test case specifies a query as a RelNode tree rather than the
usual SQL. A work around for [CALCITE-1774] was necessary to make it
work.

It turns out that [CALCITE-1769] was fixed in 1.12, probably
[CALCITE-1695].

Fix log4j.properties; now tracing goes to the console, not a file.

Close apache/calcite#440


Former-commit-id: 868d64140d446cdb33652c0c76c4038b1ca26a14