Split the mock catalog into "core" and "extended"
Not only the extended catalog take longer to create, but it creates
views, which require parsing and validation, and therefore a "simple"
validation test is no longer simple, and tends to hit a lot of spurious
break points before the real action starts happening. For most tests,
the core catalog is sufficient.


Former-commit-id: df826434d7c674caa29a0d1709a52f3e1013baf5