[CALCITE-1963] Upgrade checkstyle, and fix code to comply
Upgrade maven-checkstyle-plugin from 2.12.1 (checkstyle 5.7) to
2.17 (checkstyle 6.11.2).

Ensure compliance with the new version: lots of cosmetic changes,
especially whitespace, no functionality changes.

I would have liked to set Indentation.forceStrictCondition=true, to
ensure that parameter and argument lists are indented correctly. But it
generates lots of false-negatives in nested expressions, so we can't
do it at this point.


Former-commit-id: 9ab47c0000732ec99c3162954e1eb74eaa30cddf