[CALCITE-1998] Hive ORDER BY null values (Abbas Gadhia)
Make HiveSqlDialect aware of the version so that it may switch NULLS
FIRST/NULLS LAST behavior based on version.  Use DatabaseMetadata
major and minor version integer values instead of relying on strings
to compare versions.

Add dialect for Google BigQuery.

Change the default NullCollation for MySQL to reflect the correct collation.

Optimize the code generated for NULLS FIRST/LAST based on the order of
the sort. If the default sort would return nulls in the same order as
the NULLS FIRST/LAST flag, then we are dropping the emulation of the
NULLS FIRST/LAST clause.  Change the emulateNulls method to support
all types of NullCollations

Fix javadoc, restore MySQL calls to ISNULL. (Julian Hyde)

Close apache/calcite#545


Former-commit-id: 45425103c0be0604bb3c6c650f8f13ecb47de519