[CALCITE-2074] Simplification of point ranges that are open above or below yields wrong results
Suppose that R is a point range that is closed at one or both ends,
i.e. [x, x), or (x, x] or (x, y). Such a range is empty, but we were
wrongly simplifying it to a point. It is only valid to simplify
open-open ranges, [x, x], to a point. For example, "5 >= x AND x > 5",
described as the closed-open range "[5, 5)", should simplify to false.


Former-commit-id: 11da17b59249b4bd1bbd384a53f1cfd9ebaaf888