[CALCITE-2265] Allow comparison of ROW values (Dylan Adams)
StandardConvertletTable.consistentType now checks if structs' members
are pair-wise of the same family.

Added isSameFamily and areSameFamily methods to SqlTypeUtil to determine
whether two types, or a list of types, are of the same family. Added
unit test to verify this new functionality.

Added new case to JdbcTest to verify row comparison are executed in
HSQLDB.

Refactored common code from SqlTypeUtil.isComparable for determining
the family of a RelDataType.

Extracted SqlTypeFixture from SqlTypeFactoryTest.

Close apache/calcite#696


Former-commit-id: da568324ad84c5a792f2b9c2bd2c827c1440b19c