[CALCITE-2529] All numbers are in the same type family (Andrew Pilloud)
The linq4j code generator fails to generate calls to functions
when there is a mix between integer and floating point arguments.
This is due to the incorrect assumption that java integer types
can not be assigned to floating point types. This change fixes
the assignment mapping, enabling the generator to call floating
point functions with integer arguments.

Close apache/calcite#818


Former-commit-id: 25c332d609372522e1090cb141ab8ae5f1bdceb8