[CALCITE-2617] Add a variant of FilterProjectTransposeRule that can push down a Filter that contains correlated variables (Stamatis Zampetakis)
The default instance of FilterProjectTransposeRule does not push down
a Filter that contains correlated variables, as before.

1. Add a new constructor accepting predicates for the Filter and
   Project operands.
2. Remove the correlating condition check from the body of the rule
   and make it an optional predicate.
3. Pass the new correlating condition predicate to the existing
   constructors (retaining the same functionality for the generic
   INSTANCE rule)

Note: Classes using the protected constructor may break if they rely
on the fact that the Filter is not pushed when it contains correlating
conditions.

Add tests for each variant (Julian Hyde)

Close apache/calcite#887


Former-commit-id: 6284d3c88d9c659e0a8ef3dd5cf4819133de7522