[CALCITE-4483] WITHIN DISTINCT clause for aggregate functions (experimental)
The WITHIN DISTINCT feature is experimental and subject to
change/removal without notice.

To represent WITHIN DISTINCT in relational algebra, add
method RelBuilder.AggCall.unique() and field
AggregateCall.distinctKeys.

To represent WITHIN DISTINCT in the SQL AST, add
WITHIN_DISTINCT SQL operator.

Add a planner rule, AggregateExpandWithinDistinctRule, that
implements WITHIN DISTINCT by pre-aggregating the rows that
will go into each aggregate function.

There are two variants of the rule; one that throws a runtime
error if the operand is not functionally-dependent on the
WITHIN DISTINCT key, and one that takes the minimum. The former
is enabled by default.

The rule handles DISTINCT aggregate calls by converting them
to WITHIN DISTINCT (because the latter is a generalization of
the former).

Add an internal $THROW_UNLESS function (invoked when operands
are not functionally dependent).

Further work is required to complete the feature:
 * support FILTER clause in combination with WITHIN DISTINCT;
 * support WITHIN DISTINCT clause in GROUPING SETS queries;
 * duplicate detection fails for aggregate functions that
   respect nulls.

Regarding nulls. Luckily most aggregate functions ignore
nulls by default, but exceptions include ARRAY_AGG. If the
input to ARRAY_AGG for a given WITHIN-DISTINCT-key value is
[null, 1, null], MIN and MAX will both be 1, so we cannot
determine that all of the values were not the same.

Close apache/calcite#2339


Former-commit-id: f1da65504e598928cf77aa6a7244552692ae2529