[CALCITE-2437] FilterMultiJoinMerge should combine filters from Filter relation and multijoin post-join filters
The FilterMultiJoinMergeRule class is intended to combine a filter on
top of a multi join into a single multijoin. But instead of combining
the post join filter with the filter's condition, it simply replaces
it.

Fixing the issue by composing a conjuction of the two filters and
adding it to the new multijoin.

fixes #977


Former-commit-id: d368fc570a5d0f836c049c6acd73f2dd9f467f79