[CALCITE-2843] Support PostgreSQL cast operator '::' (Muhammad Gelbana)
SqlCastOperator is in SqlLibraryOperators but not SqlStdOperatorTable;
you can enable it using 'fun=postgresql' in your connect string.

Make SqlCastOperator a sub-class of SqlBinaryOperator to match its
syntax, delegating to SqlCastFunction for behavior such as
type-checking. Change its precedence/associativity to match that of
"::" in PostgreSQL. (Julian Hyde)

Close apache/calcite#1066


Former-commit-id: e8baf4f4720e07cdce1aa2baabf20042b9353bbb