RelOptUtilTest concurrency fixup
RelOptUtilTest was creating a static RelOptCluster which could be used
during tests running in parallel to acces the associated RelMetadataQuery
instance.  Alas the object is not safe to access from multiple threads
and in some occasions tests would fail with CyclicMetadataException.

Fixing the test class by creating a new RelOptCluster/RelBuilder
instance for each test, and removing the static instances.


Former-commit-id: 00f8b5f5ca1d45ce5029a474c42e674bec31634f