[CALCITE-3283] RelSubset does not contain its best RelNode (Xiening Dai)
In VolcanoPlanner.rename(), the given relnode will be removed when we find
there's an equivalent rel after rename. But if the node to be removed happens
to be the best relnode of its subset, the RelSubSet.best will not able to get
updated under two scenarios:
1. If equivalent rel is in the same set as the original rel, currently we do
nothing. So RelSubSet.best is not updated and points to a node that's removed
after rename.
2. If we end up merging two sub set and equivalent rel cost is same as original
rel cost, we won't update RelSubSet.best either.

Fix the issue by udpdating the RelSubSet.best if best is the RelNode that's
supposed to be removed. Enable isValid() check only under debug log level or
finer until calcite-2018 is fixed.

Close #1403


Former-commit-id: 77e7808fc0aa4e089a992535bc890047724fd900