[CALCITE-3361] Add 'lenientOperatorLookup' connection property
Add a test (redshift.iq) that parses and validates a SQL statement with
every built-in Redshift function, per the Redshift documentation.

The test uses lenientOperatorLookup so that unknown functions do not
cause validation errors (as long as they have ordinary function syntax).
Test cases are commented out if the feature is not supported (e.g. the
APPROXIMATE keyword). If a function is built into Redshift but not
known to Calcite, its test will use '!explain-validated-on' rather than
'!ok'.

In parser, increase the lookahead for JoinType() so that we can parse
LEFT SEMI JOIN. This allows us to enable BabelQuidemTest, which was
previously disabled.


Former-commit-id: 35c5f57d7db2b4745086178602ca69195ffb858e