[CALCITE-3323] Add mode to SqlValidator that treats statements as valid if they contain unknown functions (Ryan Fu)
The mode is controlled by new method SqlValidator.isLenientOperatorLookup().

If it encounters an unknown function, or a known function with invalid
number or types of arguments, the validator now creates a
SqlUnresolvedFunction. The return type is inferred to be UNKNOWN, so
that an enclosing function call can be validated.

Removed .factorypath files and updated gitignore files to reject these.

Close apache/calcite#1471


Former-commit-id: ed692fd4f2f00261348ecac779cb1a6d544d052e