[CALCITE-3441] Remove SqlTypeExplicitPrecedenceList.COMPACT_NUMERIC_TYPES because the NULL delimiters are useless
We do not need the "null" value as delimiter for the equivalence class
because method Util.skip already did that.

This also tweak the "needToCast" check for implicit type coercion:
change the "canCast" check into an assert, that means, we should not
swallow the cast when the coercion is not allowed in the type coercion matrix.


Former-commit-id: d287b0551a189fd1bf1f8a1ae3afb1b446823350