[CALCITE-2935] Support BOOL_AND, BOOL_OR, LOGICAL_AND, LOGICAL_OR aggregate functions (ShuMingLi)
BOOL_AND and BOOL_OR are in PostgreSQL and Redshift;
LOGICAL_AND and LOGICAL_OR are in BigQuery;
all are equivalent to EVERY and SOME,
and in turn to MIN and MAX over BOOLEAN types.

Rework to use SqlMinMaxAggregateFunction, and add
LOGICAL_AND and LOGICAL_OR (Julian Hyde).

Close apache/calcite#1694


Former-commit-id: b598ff08cc8ba4f00ddef08fdaf77e4fbb8d972f