[CALCITE-3751] JDBC adapter generates SQL with wrong aliases in GROUP BY ... ORDER BY query
In this particular query, RelToSqlConverter would generate something like

  SELECT x, SUM(y) FROM (
    SELECT x, SUM(y)
    FROM t
    GROUP BY x)
  ORDER BY 2

Note that 'SUM(y)' in the outer query is invalid; we should have
generated an alias for the 'SUM(y)' in the inner query, and
referred to that.

The problem was a 'ordinalMap', that maps columns called
'EXPR$n' to corresponding SqlNode. This map was global to the
RelToSqlConverter, so sometimes expressions would be used in
the wrong context.

I removed 'ordinalMap' and added a flag 'anon' which says whether
a consumer (e.g. INSERT or UNION or an enclosing SELECT) cares
about column names. If not, we remove trivial aliases like 'EXPR$0',
to make the SQL less noisy.


Former-commit-id: 3c9e156aea4a246318e1fa9ea299adfc9479e20e