[CALCITE-3756] RelSubset should not match operand(RelNode.class)
Even RelSubset is a subclass of RelNode, but they are quite different. RelNode
already fires these rules, it is not necessary to fire for RelSubset. More
importantly, it never worked due to CALCITE-3755.  In addition, it will waste
time to try to iterate class operands of RelSubset and try to match pattern
operand(RelNode), given that it never matches when ascending.

Also fixed https://issues.apache.org/jira/browse/CALCITE-3755


Former-commit-id: 0b97f45752f111b56aa499b037077d27db82ec47