[CALCITE-3807] checkForSatisfiedConverters() is unnecessary
When VolcanoPlanner registers an abstract converter, it adds the converter into set.abstractConverters list, then calls checkSatisfiedConverter() to see if any converter is satisfied and can be remove from the list. But for every abstract converter, it always satisfies itself (changeTraitsUsingConverters() returns itself). Basically the converter would be removed from the list right after it's added. So this check is completely unnecessary and it slows down the planner.


Former-commit-id: e1e5898f8975f2478862973c19799c8be45b8d83