[CALCITE-3812] Refactor HintStrategy to HintPredicate
The HintStrategy was introduced in commit
https://github.com/apache/calcite/commit/215a9b4d83e76942007fb59f46cd785e62af9ec1 [formerly 52bc246d5e07510689635f407f4d9dc494b0af4d],
at that moment, we can only define the hint match rules for propagation,
and the class name is okey.

But with more components added to the HintStrategyTable(i.e. the
HintOptionChecker, the error hander, the excluded rules), i found the
"HintStrategy" is mis-leading.

So i propose to rename HintStrategy to HintPredicate and
HintStrategyTable.Entry to HintStrategy.


Former-commit-id: 2515e9ff12a588a590f408b3660a96c3fa76cc01